% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{as_argument}
\alias{as_argument}
\title{Build list of model meta-data}
\usage{
as_argument(leadrboard)
}
\arguments{
\item{leadrboard}{the leaderboard tibble, or a filtered verison of it
from \code{\link{board}}}
}
\value{
a named list of model meta-data that can be based to a function
that builds a caret \code{train} model
}
\description{
Extracts the model meta-data from \code{\link{board}} into a list.
This data can be used to exactly reproduce the model. Can be passed
to \code{\link{run}}.
}
\examples{
parameters <- board() \%>\%
  filter(id == 1) \%>\%
  as_argument()

run(modeler, data, parameters)

}
