% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{at_last}
\alias{at_last}
\title{Get the id of the last model ran}
\usage{
at_last(number = 1)
}
\arguments{
\item{number}{number that indicates the last n models to peak at.}
}
\value{
model id of last models
}
\description{
This function returns the model ids for the last \code{number} models saved
to leaderboard. This function is intended to be used with \code{\link{peak}}.
}
\examples{
# peak at previous model
board() \%>\%
  peak(at_last())

# peak at last two models
board() \%>\%
  peak(at_last(2))

# peak at last three models
# and model 1
board() \%>\%
  peak(at_last(3), 1)

board() \%>\%
  dplyr::filter(id \%in\% at_last(3))

}
