% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{A tibble leaderboard for \code{caret} \code{train} objects}
\usage{
board(model = NULL, spec = NULL, path = NULL, level = "one",
  save = TRUE)
}
\arguments{
\item{model}{model to add to the leaderboard. The default \code{null}
means \code{leadr::board} just returns the leaderboard tibble.}

\item{path}{globally sets the path to save models and leaderboards. By
default, the path is the project directory.}

\item{level}{name of directory that contains saved models and leaderboards.
This will be a subdirectory of the specified path. The default is \code{level}
is \code{one} which gives a directory name \code{/level_one/}.}
}
\value{
\code{tibble} containing the most updated leaderboard.
}
\description{
This function (called as \code{leadr::board}) returns and possibly
updates the model leaderboard for the project.
}
\examples{
# add caret model to leaderboard
# model <- train(...)
leadr::board(model)

# to return tibble leaderboard
leadr::board()

}
