% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SedimentationRate.R
\name{calc_SedimentationRate}
\alias{calc_SedimentationRate}
\title{CF:CS model to estimate sedimentation rate from 210Pb activity}
\usage{
calc_SedimentationRate(x, reverse = TRUE, fix_a0 = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{data.frame}} (\strong{required}):
Measured 210Pb activity at different soil depths with the following
structure:\preformatted{     | depth (cm )| activity (Bq/kg) |
     |     [ ,1]  |       [ ,2]      |
     |------------|------------------|
[1, ]|    ~~~~    |       ~~~~       |
[2, ]|    ~~~~    |       ~~~~       |
 ... |     ...    |        ...       |
[x, ]|    ~~~~    |       ~~~~       |
}}

\item{reverse}{\code{\link{logical}} (\strong{optional}):
Plot the y-axis in ascending order so that the soil surface is on top.
Defaults to \code{TRUE}.}

\item{fix_a0}{\code{\link{logical}} (\strong{optional}):
In the fitted model parameter \code{a0} is the 210Pb activity at depth \code{z=0}.
If the argument \code{fix_a0} is \code{TRUE}, this function will take the first
activity value (i.e., the value closest to the surface) as a fixed value
in the equation. For \code{fix_a0=FALSE} parameter \code{a0} is estimated by the model.
Defaults to \code{FALSE}.}

\item{verbose}{\code{\link{logical}} (\strong{optional}):}

\item{...}{Futher arguments passed to \code{\link{plot}}.}
}
\value{
A \code{\link{list}} with the following structure:

\tabular{lll}{
Element \tab Data type \tab Description \cr
\code{$summary} \tab \code{\link{data.frame}} \tab summary table of important model results \cr
\code{$data}    \tab \code{\link{data.frame}} \tab original input data \cr
\code{$model}    \tab \code{\link{nls}} \tab model output generated by \code{\link[minpack.lm:nlsLM]{minpack.lm::nlsLM}}
}
}
\description{
This function fits a "Constant Flux Constant Sedimentation Rate" (CF:CS)
model to 210Pb activity data from different soil depths
to estimate the sedimentation rate of the profile.
}
\details{
\strong{Model}

\deqn{
  y = a0 * exp((l/s) * x)
}
\itemize{
\item \code{y} = 210Pb activity at depth \code{x} (cm)
\item \code{a0} = 210Pb activity at depth \code{x=0}
\item \code{l} = 210Pb decay constant (= 0.03114 a-1)
\item \code{s} = sedimentation rate (cm/a)
\item \code{x} = depth below surface (cm)
}
}
\examples{

# Load example data (synthetic)
data(Pb)

# Apply the model
results <- calc_SedimentationRate(x = Pb,
                                  reverse = TRUE,
                                  fix_a0 = FALSE,
                                  verbose = TRUE)


}
\references{
Szmytkiewicz, A., Zalewska, T., 2014. Sediment deposition and accumulation
rates determined by sediment trap and 210Pb isotope methods in the
Outer Puck Bay (Baltic Sea). Oceanologia 56(1), 85-106.
}
\author{
Christoph Burow, University of Cologne
}
