% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{caret_seed}
\alias{caret_seed}
\title{Set seed for \code{caret}'s \code{trainControl}}
\usage{
caret_seed(seed = 1, number)
}
\arguments{
\item{seed}{starting seed for list of seeds}

\item{number}{number of folds or resamples}
}
\value{
List of seeds used by \code{trainControl}.
}
\description{
Given a starting seed and number of folds (or resamples),
this function creates a vector of seeds needed for the `caret`
trainControl object.
}
\examples{
trainControl(method = 'cv', number = 10, seeds = caret_seed(1234, 10))

}
