% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{git_ignore_models}
\alias{git_ignore_models}
\title{Add \code{model*.RDS} to .gitignore}
\usage{
git_ignore_models()
}
\description{
Adds file types generated by
\code{board} of type \code{model*.RDS} to .gitignore. Model types
can be very large, so you may not want to save them to a git repo.
}
