% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{last_model}
\alias{last_model}
\title{Get the index of the last model ran}
\usage{
last_model(num = 0)
}
\arguments{
\item{num}{used to indicate the number of models before the previously ran model.
The default \code{0} indicates the previously ran model. \code{k} would mean the
model ran \code{k} times before the previous model.}
}
\value{
Number from the leaderboard column \code{num} for the last model.
}
\description{
This function returns the \code{num} value from the leaderboard for the last
model ran. This function is intended to be used with \code{\link{peak}}.
}
\examples{
# peak at previous model
leadr::board() \%>\%
  peak(last_model())

# peak at model before previous
leadr::board() \%>\%
  peak(last_model(1))

}
