% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadR-package.R
\docType{package}
\name{leadR-package}
\alias{leadR-package}
\title{Sediment Deposition and Accumulation Rates Determined by 210Pb Isotope Methods}
\description{
A minimal package to estimate the sedimentation rate of a soil profile
using a 'Constant Flux Constant Sedimentation Rate' model applied to measured
210Pb activity.
}
\author{
Christoph Burow (University of Cologne, Germany)
}
\keyword{package}
