% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeadR.R
\name{leadR_CIC}
\alias{leadR_CIC}
\title{Pb-210 CIC models}
\usage{
leadR_CIC(filename, unsupp = "NULL", coredate, export = NULL)
}
\arguments{
\item{filename}{the filename of the CSV datafile. It must have the following columns: "Depth", the center depth of the sample (in cm), Depth_Span, the width of the sample (in cm), Pb.210, the measured activity (in Bq/g), and "Pb.210_1s" the standard error of that measurement (in Bq/g).}

\item{unsupp}{optional, if passed should be a numeric value to use as the baseline Pb-210 (in Bq/g), or "estimate" to automatically estimate it. If you don't pass either you must supply the data as a column in the input data.}

\item{coredate}{the data of the top of the core (in years AD).}

\item{export}{optional, the name to use for an output datafile.}
}
\value{
an object with the raw data, input, output, and accumulation rate
}
\description{
Performs a constant initial accumulation model for Lead-210 data.
}
\examples{
leadR_CIC(filename=system.file("extdata", "exampledata.csv", package = "LeadR"), coredate=2014)

}
