% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeadR.R
\name{leadR_CRS}
\alias{leadR_CRS}
\title{Pb-210 CRS models}
\usage{
leadR_CRS(filename, unsupp = "estimate", coredate, area = "UWITEC",
  export = NULL)
}
\arguments{
\item{filename}{the filename of the CSV datafile. It must have the following columns: "Depth", the center depth of the sample (in cm), Depth_Span, the width of the sample (in cm), Pb.210, the measured activity (in Bq/g), and "Pb.210_1s" the standard error of that measurement (in Bq/g), and "dry_mass", the dry mass of the sample (in grams).}

\item{unsupp}{optional, if passed should be a numeric value to use as the baseline Pb-210 (in Bq/g), or "estimate" to automatically estimate it. If you don't pass either you must supply the data as a column in the input data.}

\item{coredate}{the data of the top of the core (in years AD).}

\item{area}{the diameter of the sediment core, or just "UWITEC" if you're using a UWITEC corer.}

\item{export}{optional, the name to use for an output datafile.}
}
\value{
an object with the raw data, input, output, and accumulation rate
}
\description{
performs a constant rate of supply model for Lead-210 data.
}
\examples{
leadR_CRS(filename=system.file("extdata", "exampledata.csv", package = "LeadR"), coredate=2014)

}
