% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{model_list}
\alias{model_list}
\title{Convert the leaderboard tibble to a list of models}
\usage{
model_list(leadrboard)
}
\arguments{
\item{leadrboard}{the leaderboard tibble, or a filtered verison of it
from \code{\link{board}}}
}
\value{
a named list of caret models (\code{train} objects)
}
\description{
Given a possibly filtered leaderboard tibble from \code{\link{board}},
\code{model_list} returns every model in a list. Each entry in the
list in named by the model method and id.
}
\examples{
model_list <- board() \%>\%
  filter(group == 1) \%>\%
  model_list()

}
