% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{modeler}
\alias{modeler}
\title{Wrapper around caret \code{train}}
\usage{
modeler(data, method = "cv", num = 5, index = NULL, seeds = NA, model,
  tune = NULL)
}
\value{
a caret \code{train} object
}
\description{
This is a wrapper around caret \code{train} that accepts the
model meta-data from \code{\link{board}}. This function is used
to exactly reproduce models in the leader board, or run
new models.
}
\details{
The parameters correspond to the arguments in caret's
\code{trainControl} and \code{train} functions.

The source code for this function is also an example how to
create your own modeler function for your own needs.

Note: you must manually load the caret package to use this function.
caret is not a dependency of leadr.
}
\examples{
library(caret)
library(leadr)
modeler(iris, model = "rf")

}
