% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{oof_grab}
\alias{oof_grab}
\title{Return out of fold model predictions}
\usage{
oof_grab(models, train_data, type = "raw")
}
\arguments{
\item{seed}{starting seed for list of seeds}

\item{number}{number of folds or resamples}
}
\value{
tibble or list of tibbles with out of fold predictions
for each model.
}
\description{
Given a model or list of models, this function returns the
out of fold predictions. These out of fold predictions can be used
to make stacked or blended ensembles. See this vignette for more
details.
}
\examples{
trainControl(method = 'cv', number = 10, seeds = caret_seed(1234, 10))

}
