% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadR-package.R
\docType{data}
\name{Pb}
\alias{Pb}
\title{Synthetic 210Pb Activity Data}
\format{A \code{\link{data.frame}} with two with two columns:
\tabular{ll}{
\code{$depth} \tab depth below surface (cm) \cr
\code{$activity} \tab 210Pb activity (Bq/kg) \cr
}}
\description{
A synthetic data set of 210Pb activity per units of depth below surface. See
example on how the data was generated.
}
\examples{
## How the data was generated
# a0 = 100 Bq/kg
# s = 0.5 cm/a
Pb <- data.frame(depth = seq(0, 30, length.out = 20),
                 activity = jitter(
                    100 * exp(-0.03114 / 0.5 * seq(0, 30, length.out = 20)),
                    factor = 15))

## Use example data
data(Pb, envir = environment())
r <- calc_SedimentationRate(x = Pb)

}
\keyword{datasets}
