% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{peak}
\alias{peak}
\title{Peak at the leaderboard position of a model}
\usage{
peak(leadrboard, num)
}
\arguments{
\item{leadrboard}{leadrboard tibble returned by \code{\link{board}}}

\item{num}{the number of the model. See \code{\link{last_model}} to easily
specify model number.}
}
\value{
A filtered leadrboard tibble with only 9 rows, centerd around the
supplied model number.
}
\description{
This function returns a tibble with the 8 models ranked closest to the
supplied model in accuracy. Usually chained with the magrittr pipe \code{\%>\%}.
}
\examples{
leadr::board() \%>\%
  peak(last_model())

}
