% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{run}
\alias{run}
\title{Run function taking the model meta-data}
\usage{
run(modeler, formula, data, parameters)
}
\arguments{
\item{modeler}{user created function that is a wrapper around
caret \code{train}}

\item{formula}{formula passed to caret \code{train} function}

\item{data}{training data for the caret \code{train} function}

\item{parameters}{parameters that go into the modeler wrapper. If
a previous model is being rerun, use \code{\link{as_argument}}
to extract model meta-data from the \code{\link{board}}}
}
\value{
returns the return value of the \code{modeler} function
supplied, which should be a caret \code{train} object.
}
\description{
Runs a function that takes a data to train on and a
list of parameters. This is a wrapper around \code{\link{do.call}}.
}
\examples{
parameters <- board() \%>\%
  filter(id == 1) \%>\%
  as_argument()

run(modeler, Species ~ ., iris, parameters)

}
