% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{A tibble leaderboard for \code{caret} \code{train} objects}
\usage{
board(model = NULL, path = here::here("models"), dir = "initial",
  save = TRUE, quiet = FALSE)
}
\arguments{
\item{model}{model to add to the leaderboard. If no model is supplied
(the default \code{null}), \code{board} returns the leaderboard tibble
for the project.}

\item{path}{the path to the saved model directory. By
default, the path is the project root. For best results, the path
string should be constructed with \code{file.path} or
\href{https://github.com/krlmlr/here}{\code{here::here()}}.}

\item{dir}{the name of directory where the model is saved.
This will be a subdirectory of the specified path. The default directory
is named \code{initial}. This is commonly used to group similar models.}

\item{save}{whether \code{board} should save the supplied model to \code{dir}. If
\code{FALSE} the model will not be saved, but will be added to the leaderboard.}

\item{quiet}{whether \code{board} should return the leaderboard tibble to the console.
By default \code{quiet = FALSE} means that the tibble prints to console.
\code{quiet = TRUE} is useful in a \code{.Rmd} environment, where you want to add
the model to the leaderboard without printing the tibble.}
}
\value{
\code{tibble} containing the most up-to-date leaderboard.
}
\description{
This function updates and returns
the model leaderboard for the project. Please read the
\href{https://github.com/tmastny/leadr}{README} and the
introduction
\href{https://tmastny.github.io/leadr/articles/introduction.html}{vignette}.
}
\examples{
# add caret model to leaderboard
# model saved to "file.path("models", "initial")"
model <- train(...)
board(model)

# return tibble leaderboard
board()

# save to different directory
ensemble <- train(...)
board(ensemble, dir = "ensembles")

}
