% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_tools.R
\name{oof_grab}
\alias{oof_grab}
\title{Return out of fold model predictions}
\usage{
oof_grab(models, type = "raw")
}
\arguments{
\item{models}{A model or list of models to get the predictions}

\item{type}{the results of the prediction. For classification models,
\code{"raw"} returns the outcome label and \code{"prob"} returns the
label probabilities.}
}
\value{
a tibble with one column per model and a column of the training data
outcomes. If \code{type = "prob"} there will be n columns per model, where
n is the number of labels in the outcome.
}
\description{
Given a model or list of models, this function returns the
out of fold predictions. These out of fold predictions can be used
to make stacked or blended ensembles. See the ensemble
\href{https://tmastny.github.io/leadr/articles/ensemble.html}{vignette}
for examples.
}
\examples{
oofs <- oof_grab(models)

}
